JDF 2.13.0-0, part2
- updated extensions

+++ full/profile/extensions/{b9db16a4-6edc-47ec-a1f4-b86292ed211d}/resources/downloadhelper/lib/chunks.js
@@ -437,7 +437,7 @@ var ChunkStopRecordAction = merge(Class({
 	removeChunkSet: function(chunkSet,removeHit) {
 		Cu.import("resource://gre/modules/FileUtils.jsm");
 		var file =  new FileUtils.File(chunkSet.dir);
-		if(file.exists)
+		if(file.exists && !simplePrefs.prefs['chunk.keep-temp'])
 			file.remove(true);
 		if(removeHit)
 			hits.remove(chunkSet.hit.id);
